requirements.txt
aiohttp
beautifulsoup4
aiosqlite
pyyaml
feedparser
apscheduler
python-dotenv
trafilatura
lxml
tqdm
PyQt5
pytest
pytest-asyncio
pytest-cov
playwright
pytest-playwright
